+++ CHANGELOG.md
@@ -5,7 +5,13 @@
 ### Added
 
 * `tools` command for tool management. ([#27] by [@xStrom])
+* Automatic installation of tools from source via `cargo`.
+  Now implemented for `ripgrep`. ([#28] by [@xStrom])
 * `--strict` option to `clippy`, `copyright`, and `format` commands to use locked tool versions. ([#27] by [@xStrom])
+
+### Changed
+
+* `ci` command now also runs the `copyright` command. ([#28] by [@xStrom])
 
 ## [0.2.0] - 2026-02-07
 
@@ -39,6 +45,7 @@
 [#23]: https://github.com/Nevermore/prep/pull/23
 [#24]: https://github.com/Nevermore/prep/pull/24
 [#27]: https://github.com/Nevermore/prep/pull/27
+[#28]: https://github.com/Nevermore/prep/pull/28
 
 [Unreleased]: https://github.com/Nevermore/prep/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/Nevermore/prep/compare/v0.1.0...v0.2.0